convert applyEach tests to mocha
